requirements.txt
absl-py==1.2.0
astunparse==1.6.3
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
dill==0.3.5.1
etils==0.7.1
flatbuffers==1.12
gast==0.4.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio @ file:///Users/runner/miniforge3/conda-bld/grpc-split_1660380895008/work
h5py @ file:///Users/runner/miniforge3/conda-bld/h5py_1637964070496/work
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1653326072482/work
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
promise==2.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1660317639764/work
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.15.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.6.0
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.2
tensorflow-metadata==1.9.0
tensorflow-metal==0.5.0
termcolor==1.1.0
toml==0.10.2
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
